+++ BE/DotNet_WEB/Module/chuc_nang/chuc_nang_trang_cong_ty/chuc_nang_phuc_loi_cong_ty/chuc_nang_phuc_loi_cong_ty_web.cs
@@ -23,6 +23,10 @@ namespace DotNet_WEB.Module.chuc_nang.chuc_nang_trang_cong_ty.chuc_nang_phuc_loi
         {
             using var coon = new MySqlConnection(chuoi_KetNoi);
             coon.Open();
+            if (pl == null || pl.phuc_Loi_Cong_Ty == null || pl.phuc_Loi_Cong_Ty.Count == 0)
+            {
+                return false;
+            }
             string sql = @"INSERT INTO phuc_loi_cong_ty (ma_cong_ty, ten_phuc_loi, mo_ta)
             VALUES (@ma_cong_ty, @ten_phuc_loi, @mo_ta)";
             int r = 0;
@@ -38,14 +42,17 @@ namespace DotNet_WEB.Module.chuc_nang.chuc_nang_trang_cong_ty.chuc_nang_phuc_loi
             }
             return r > 0;
         }
-        
+
         public static bool xoaPhucLoiCongTy(phuc_loi_cong_ty_cap_nhat pl)
         {
+            if (pl == null || pl.phuc_Loi_Cong_Ty == null || pl.phuc_Loi_Cong_Ty.Count == 0)
+                return false;
+                
             using var coon = new MySqlConnection(chuoi_KetNoi);
             coon.Open();
             string sql = @"delete from phuc_loi_cong_ty where ma_cong_ty = @ma_cong_ty and ma_phuc_loi_cty = @ma_phuc_loi_cong_ty";
             int r = 0;
-            foreach(var plct in pl.phuc_Loi_Cong_Ty)
+            foreach (var plct in pl.phuc_Loi_Cong_Ty)
             {
                 using (var cmd = new MySqlCommand(sql, coon))
                 {
